bpo-32904: Fix a potential crash in os.chdir() on Windows
win32_wchdir() retries GetCurrentDirectory() with a larger buffer
if necessary, but doesn't check whether the new buffer is large enough.
Another thread could change the current directory in meanwhile,
so the buffer could turn out to be still not large enough, left in
an uninitialized state and passed to SetEnvironmentVariable() afterwards.

While win32_wchdir() doesn't guarantee a consistent result in presence
of races, crashing is probably a bad idea.